refactor: preserve commit detail fast refresh
